pyproject.toml
[tool.poetry]
name = "ublue-scanner"
version = "0.1.0"
description = ""
authors = ["John Anderson <sontek@gmail.com>"]
readme = "README.md"
packages = [{include = "ublue_scanner"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.7.0"
python-dotenv = "^1.0.0"
retry = "^0.9.2"
ghapi = "^1.0.3"
arrow = "^1.2.3"
rich = "^13.3.2"
ruamel-yaml = "^0.17.21"
httpx = "^0.23.3"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ublue-scan = "ublue_scanner.cli:app"
